Bean Validation

+++ src/main/java/com/mayara/e_commerce/services/ProductService.java
@@ -1,12 +1,16 @@
 package com.mayara.e_commerce.services;
-
 import com.mayara.e_commerce.dtos.ProductDTO;
 import com.mayara.e_commerce.entities.Product;
 import com.mayara.e_commerce.repositories.ProductRepository;
+import com.mayara.e_commerce.services.exceptions.DatabaseException;
+import com.mayara.e_commerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -17,7 +21,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(
+                ()->  new ResourceNotFoundException("não encontrado"));
         return new ProductDTO(product);
     }
     @Transactional(readOnly = true)
@@ -36,17 +41,30 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id,ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(entity,dto);
-        repository.save(entity);
-        return new ProductDTO(entity);
+        try{
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(entity,dto);
+            repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Não encontrado");
+        }
+
     }
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw  new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
+
     }
 
-    private void copyDtoToEntity(Product entity, ProductDTO dto) {
+    private void copyDtoToEntity(Product entity, ProductDTO dto){
         entity.setName(dto.getName());
         entity.setDescription(dto.getDescription());
         entity.setPrice(dto.getPrice());
